requirements.txt
Flask==2.0.3
python-telegram-bot==13.7
spotipy==2.19.0
requests==2.26.0
python-dotenv==0.19.2